Collision Training Observations refactored
Collision Training Observation will now train more efficiently for training with more than to Bot Vehicles on the track.

The brain is now given the distance to the closest car as an input rather than having to determine which car is closest by itself given the input of all cars on the track.

This will also be reusable for instances of different number of Bot vehicles being on the track.