Added 3-2-1-GO! to start of race
Created a script utilsing coroutines and lerps to animate a 3-2-1-GO! countdown in front of the camera as a race begins. Script has been added to the main camera. You can also easily edit the colors of the numbers/GO! as well as the time between numbers and a buffer before the countdown begins via the serialized fields.