ML Training Game Restart Vehicle Respawn Bug Fix
During training the game is set to restart from the map origin after a certain number of episodes to avoid the track approaching the maximum positional values and crashing the game.

However, when this reset occurred, the bot vehicles were not being re-positioned back to the map origin with the track pieces.